libgbm: make possible to use other libgbm provides than mesa
Mesa is the default libgbm provider, but does not yet have virtual/libgbm
in the PROVIDES list. This makes it difficult the change the dependency
to another recipes that provide only libgbm. Add virtual/libgbm to mesa
and TI's libgbm recipe, and change the dependency in qtbase.

Add gbm to qtbase PACKAGECONFIG on intel devices.

Task-number: QTBUG-57896
Change-Id: I491df7057e52f9e74dd87abf65e542eab2cd3f9e
Reviewed-by: Gatis Paeglis <gatis.paeglis@qt.io>
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>