qdbd: Add network configuration script
The script is called to reset the host-device network interface before
starting qdbd. qdbd will call it to configure the network when it is
told by the host what network to use.

Due to dependency to systemd-networkd in the script and the overall
system, working with SysV init is fiction, so remove that configuration.

Change-Id: I86f8348e57077c8718cedf68af82796adef65f07
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>