Fix DNS for Qemu running on b2qt toolchain sysroot
The b2qt toolchain had broken link in sysroot/etc/resolv.conf. It was
pointing to ../run/resolv.conf which is correct on actual device that
gets the file dynamically from systemd. That file is never created for
the toolchain and DNS operations fail. This change removes the link
from the toolchain.

Task-number:  QTBUG-62580
Change-Id: I52098e667a9190e300a06b777f9fac778a545faf
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>