emulator: fix conflicting glesstream packages
qtglesstream-dummy-client is needed to provide qtbase with opengl headers
and dummy library and the real qtglesstream is then built against qtbase
to provide it actual implementation. Now qtglesstream-dummy-client is built
to provide files to the sysroot, but not any packages, and qtglesstream is
built to provide the packages, but to deploy nothing to the sysroot.

Solves do_rootfs error: package packagegroup-b2qt-embedded-base-1.0-r0.all
requires qtglesstream, but none of the providers can be installed

Change-Id: Ib388599d9240a5ab9351692a4509adf721cc995c
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>