fix binutils-cross rebuild issue with mingw
Upstream poky recipe incorrectly empties LDGOLD value for binutils-cross
based on the sdkmingw32 override. This causes all the target packages to
be build twice: first when SDKMACHINE is empty, and second when time
when SDKMACHINE is set to mingw32.

The actual value use here must be exactly what is used in
meta/recipes-devtools/binutils/binutils.inc

Change-Id: Id6107bae78cf57669b04cea6467dcf11d1f5c913
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>