qtivi: Fix the broken python3 binary inside an yocto SDK
QtIvi creates a virtualenv using the python3 binary from the native
sysroot. This virtualenv is also used for the SDK. As it is not intended
for the SDK it is using the wrong interpreter location.

This workaround installs the python3 binary into the nativesdk sysroot
where the qtivi nativesdk recipe can pick it up and copy it into its
virtualenv.

This needs to be improved in future versions and is discussed here:
AUTOSUITE-176

Task-number: AUTOSUITE-179
Change-Id: Ib134231204cc8c6a56f2a93bedb553ecbcbd9123
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>
Reviewed-by: Tobias Olausson <tobias.olausson@pelagicore.com>