Add QMLPreview and QtQuickTimeline modules
The Timeline and the QML live preview are introduced as part of the
Qt Quick Designer NG project/initiative.

Since the sources are not available publicly, modules are included in
the image only when building internally in Qt Company network, where
it's assumed that user has access to the private repos in codereview.

Task-number: QTBUG-67140
Change-Id: I822d527ff132e89a8c3205e267a88c60c50fb387
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>