tegra: include tegra flasher utils into release image
Since the sd card image won't bootup on a device which has been flashed
with different version of L4T, include flasher utils that can be used to
flash the image into the device as well.

The new prepare-image.sh script will use the SD card image to create
an image that can be flashed into the device. This saves user from
downloading matching image from NVIDIA or bitbaking a flashable image.

Change-Id: I28cdcd10a531eaf9fd511565e2542bd0d81541b6
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>
Reviewed-by: Kari Oikarinen <kari.oikarinen@qt.io>