intel: rework initramfs support
The WKS image creation ignored our old initramfs. Change to using
an existing core-image-minimal-initramfs recipe and extend it with
support for the internal MMC reader. Add separate recipe to combine
the intel microcode and the new initramfs, and add wks file to use it.
Remove all old recipes that are no longer used.

Change-Id: I4e1f244685a7f2bd304aa23d9e07580fa9dc9652
Reviewed-by: Timo Aarnipuro <timo.aarnipuro@qt.io>
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>