Add QtSafeRenderer 1.1
Add recipe for the QtSafeRenderer and include it in the automotive
image and toolchain. The sources are in a private git repository,
so it's included in the build only if we know that we have access
to those (internal-build) or user has explicitly enabled it in the
local.conf (DISTRO_FEATURES_BACKFILL). The recipe also knows how to
use the sources from a Qt online SDK intallation if the variable
(QT_SDK_PATH) is setup correctly.

Neptune3 is enabled to use QSR if that is available.

Task-number: AUTOSUITE-1009
Change-Id: I2efc68a598f1e83324de40f4f0cccce9615045f7
Reviewed-by: Nikolay Zamotaev <nzamotaev@luxoft.com>
Reviewed-by: Jani Launonen <jani.launonen@qt.io>