jetson: Use default KERNEL_ROOTSPEC for jetson boards
Upstream meta layer has support for specifying different
partition to boot. This relies on environment variable
that the default bootloader doesn't have, which meant
that SD card boot would not work unless image is first
flashes to the device. Revert back to the default value,
so that sd card boots work again.

Fixes: QTBUG-75091
Change-Id: I8c519d65bd6201246dc2ca1e4db663bc21f773b8
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>