Add support external boot2qt bsp layers
Change the initialization script so that it can used with additional
boot2qt bsp layer that can bring new hardware support to Boot2Qt
without need to update these scripts for every device.

New commandline options are added to support this use case:
--group options allows specifying which meta layers to use without having
to add the specific device to the script.
--bsp option is used make sure that the separate boot2qt bsp layer is
correctly added to the sources/ directory as a symbolic link after
repo-tool is used to clone it.
--manifest option allows specifying the used manifest file if the
name or location differs from the defaults.

Task-number: QTBUG-73935
Change-Id: Ibcbe70bbd3c7bc69954178e0dfe25dccc6dd5e8e
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>