Support using setup scripts with manifest repo
Prepare scripts for use with separate manifest repo, allowing full
build environment to be initialized with repo tool without need to
first clone meta-boot2qt.

Mirroring is now removed, as it makes more sense to use it with the
manifest repo as it then mirrors also meta-boot2qt repo.

The bblayers.conf and local.conf samples are now searched from
'sources/templates' directory, which is setup to point to either
meta-boot2qt or any boot2qt bsp layer that might be used. Same setup
is also done in the manifest repo without need to run the
b2qt-init-build-env script. This also allows setup-environment.sh
to work without the --bsp argument.

Change-Id: I9b8bfaf5d39ca2cda4cff65f196338ac96a6ba31
Reviewed-by: Timo Aarnipuro <timo.aarnipuro@qt.io>
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>