toradex: fix missing qmake during tezi image creation
The teziimg task is using prefuncs which are executed before the
task dependencies are resolved, which means that qmake might not
be available in native-sysroot. This does not fail always, as the
do_image_conf task also has dependency to native-qtbase and both
image tasks are executed in parallel.

Move the dependency earlier to make sure qmake is always correctly
available.

Change-Id: I68af7e70f4458d5e0eb01c893b5546672db719a9
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>