tegra: update to L4T 32.3.1
Starting with L4T R32.3.1 and JetPack 4.3, The Tegra Multimedia API
kit has moved to JetPack, so all builds now require you to set up
an SDK Manager downloads area or use of the The Qt company internal
playground git repository.

NVIDIA_DEVNET_MIRROR is set to the internal repository and needs to
be changed (e.g. in local.conf) to your local path if not using the
playground repository.

Booting from SD card won't work unless the eMMC has the matching
u-boot and/or kernel.

Change-Id: Id892ed6a6b367b24bd84f9225ffbbb94fb4a1513
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>