nxp: default to NXP bsp
By default, meta-freescale uses mainline bsp for all imx6 boards, which
in turn will use kms backend from mesa for the graphic integration in Qt.
The KMS backend still has some issues, so instead switch to using the
vivante proprietare graphics and backend.

Change the recommended vulkan drivers to imx specific, so that we don't
unnecessarily build mesa when it's not needed, when it in fact doesn't
even build.

Task-number: QTBUG-85858
Change-Id: I47d470907ca946213f9dd628e965ff0ffc09b680
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>