default-qt-envs: add wayland display env for weston
Weston is now running as 'weston' user and using /run/wayland-0 as
the default socket for connecting to the compositor. Make sure that
applications can connect to it by adding environment variable in
the default envs.

Add also QT_QPA_PLATFORM to make eglfs or linuxfb the default platform.
If not added, Qt would default to wayland since it finds WAYLAND_DISPLAY
from the environment.

Change-Id: Ic8581ad0ee486aa27c1c36c2f89aceb76dbcb387
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>