Split the image types bbclass
Split the image types bbclass so that each device has now its own
bbclass with the overrides that they need. We are only left with
tezi overrides for Toradex boards and tegra overrides for Jetson
boards. The extra space addition is now applied only in the boot2qt
demo image, so custom images are not affected.

Fix dependency in deploy-buildinfo which was hidden previously
since all machines inherit the image types bbclass.

Pick-to: 6.2
Change-Id: I7c9cbe014b7f51cc31885d6b4da62da0bb9622fd
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>