sdk: remove compiler configs from creator integration
We cannot set compilers to CMake configs, since the actual compilers
settings in the toolchain file contains extra arguments that are needed
for the toolchain to work correctly.

This will again cause QtCreator to show the warnings for missing
configurations for the Kit, but there's nothing we can do about that.

Task-number: QTBUG-100123
Pick-to: 6.3 6.2 6.2.3
Change-Id: I963074dd971d651f1de6af93a19d57fd2fa42628
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>