swupdate: add var-image-swupdate-b2qt recipe
Add recipe to create an swupdate image based on b2qt.
Since most of the b2qt packages were included in the
b2qt-embedded-qt6-image.bbappend, it was not possible
to require it in the var-image-swupdate-b2qt recipe.
Those packages were moved to var-b2qt.inc to be included
in both b2qt and swupdate recipes.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>