conf: distro: Remove ld-is-gold from DISTRO_FEATURES
When using B2Qt SDK toolchain, Linux kernel build exits
with gold linker error.

Error message:
gold linker is not supported as it is not capable of linking the kernel proper.
scripts/Kconfig.include:56: Sorry, this linker is not supported.

This regression introduced by the commit that adds
ld-is-gold to DISTRO_FEATURES in b2qt.conf
link: https://github.com/varigit/meta-variscite-boot2qt/commit/84eeee07b2bb88179e38dbb5aa1f982fcbbad278

Removing ld-is-gold from DISTRO_FEATURES fixes this issue.

Signed-off-by: Natalia Kovalenko <natasha.k@variscite.com>